requirements.txt
blinker==1.6.2
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.4
distlib==0.3.7
exceptiongroup==1.1.2
filelock==3.12.2
Flask==2.3.2
identify==2.5.25
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
jinaai==0.2.1
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.1
platformdirs==3.9.1
pluggy==1.2.0
pre-commit==3.3.3
pydantic==1.10.11
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
replicate==0.8.4
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
virtualenv==20.24.1
Werkzeug==2.3.6
